refactor(messaging): moved `reply` function to dedicated folder
